Seperate drops and lost
Drops L12
Lost L13

This is so that we can identify if it's a local overflow or delivery
problems.

This also fixes a problem where the "since" part of the error messages
are incorrect. Previously it just used time.Now(). Now we update a
timestamp every time the counter is ReadAndReset(). This timestamp is
written out instead.