Handle Content-Length, Drops, Lost again
Re: Content-Length, this is only until (hopefully) logplex supports
chunked encoding. It's expensive to go through all of the lines and
re-make formatters for them to make sure we cna get the right Length().

Re: Drops/Lost. I consider this a first step, there is a bunch of
logplex stuff in outlet still, but I think this is a step in the right
direction towards something more generic.