Cleans up the formatters a little bit
HTTPFormatter's Request() returns a *http.Request that is ready for the
outlet to post. The outlet may mark up that request if so desired, but
it's otherwise ready to be handled by the http.Client.

SubFormatter doesn't implement Request because it doesn't make any
sense for them.

A bunch of other misc cleanups