[kinesis-sharding] Added kinesis shards flag
By letting the user specify the number of shards he/she will use on Kinesis, we can make the cardinality
of the partition keys we create match the shard count. With the number of partition keys equal to the
number of Kinesis shards, we make sure there is a good chance that the records we are sending to Kinesis
are distributed equally between shards.