Initial support for sending logs to Amazon's CloudWatchLogs service
To send logs to AWS CloudWatchLogs provide a logs-url formatted like so:
https://logs.<region>.amazonaws.com/<log group name>/<log stream name>

For more information on log groups & streams see:
https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/Working-with-log-groups-and-streams.html

Posting to an AWS Cloud Watch Log stream returns a token, which must
be provided to the next post. log-shuttle reads the current token on
startup to get an initial value. log-shuttle passes the token returned
from the last post to the next post. This mean that no matter how many
outlets are configured only a single post to a cloudwatch log stream
will happen at once.

TODOS:
* more robust unit and integation tests
* ensure recovery of ability to stream logs after errors as the token
  may get lost/dropped